refactor helper functions related to chats in a file
